@@ -5,26 +5,27 @@ version = "0.5.1"
 authors = ["Anselm Eickhoff <anselm.eickhoff@gmail.com>"]
 repository = "https://github.com/aeickhoff/kay"
 license = "MIT"
-edition = "2018"
+edition = "2024"
+
 [badges]
 maintenance = { status = "experimental" }
 
 [dependencies]
 byteorder = "1"
 chunky = "0.3.7"
-compact = "0.2.13"
+compact = {git = "https://github.com/hsiang-lee/compact.git", rev = "5018c65c9a"}
 compact_macros = "0.1.0"
-url ="1.7.2"
+url ="2.5.7"
 serde = {version = "1.0", optional = true}
 serde_derive = {version = "1.0", optional = true}
 
 [dependencies.tungstenite]
-version = "0.5.3"
+version = "0.27.0"
 optional = true
-default-features = false
+features = ["handshake"]
 
 [dependencies.stdweb]
-version = "0.4.7"
+version = "0.4.20"
 optional = true
 
 [features]
