@@ -85,9 +85,9 @@ impl<T> Compact for External<T> {
         0
     }
     unsafe fn compact(source: *mut Self, dest: *mut Self, _new_dynamic_part: *mut u8) {
-        ::std::ptr::copy_nonoverlapping(source, dest, 1)
+        unsafe { ::std::ptr::copy_nonoverlapping(source, dest, 1) }
     }
     unsafe fn decompact(source: *const Self) -> Self {
-        ::std::ptr::read(source)
+        unsafe { ::std::ptr::read(source) }
     }
 }
